+++ src/test/java/coffee/synyx/auth/authentication/web/LogoutControllerIT.java
@@ -11,7 +11,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 
 import org.springframework.security.test.context.support.WithMockUser;
 
-import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
@@ -26,15 +25,16 @@ import static org.springframework.security.test.web.servlet.setup.SecurityMockMv
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.redirectedUrl;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;
 
 
 /**
+ * @author  Yannic Klem - klem@synyx.de
  * @author  Tobias Schneider - schneider@synyx.de
  */
 @RunWith(SpringRunner.class)
 @SpringBootTest
-@TestPropertySource(locations = "classpath:application-test.properties")
-public class RootControllerTest {
+public class LogoutControllerIT {
 
     @Autowired
     private WebApplicationContext webContext;
@@ -47,27 +47,35 @@ public class RootControllerTest {
     @Before
     public void setupMockMvc() {
 
-        mockMvc = MockMvcBuilders.webAppContextSetup(webContext)
-                .apply(springSecurity(springSecurityFilterChain))
-                .build();
-    }
-
-
-    @Test
-    public void redirectToRootIfNotLoggedIn() throws Exception {
-
-        ResultActions resultActions = mockMvc.perform(get("/"));
-        resultActions.andExpect(status().is3xxRedirection());
-        resultActions.andExpect(redirectedUrl("http://localhost/login"));
+        mockMvc = MockMvcBuilders.webAppContextSetup(webContext).apply(springSecurity(springSecurityFilterChain))
+            .build();
     }
 
 
     @Test
     @WithMockUser
-    public void redirectToRootIfLoggedIn() throws Exception {
+    public void returnsLogoutPageIfLoggedIn() throws Exception {
 
-        ResultActions resultActions = mockMvc.perform(get("/"));
-        resultActions.andExpect(status().is3xxRedirection());
+        ResultActions resultActions = mockMvc.perform(get("/logout"));
+        resultActions.andExpect(status().isOk());
+        resultActions.andExpect(view().name("auth/logout"));
+    }
+
+
+    @Test
+    public void redirectsToDefaultRedirectUriIfNotLoggedInAndNoReferrerHeaderExists() throws Exception {
+
+        ResultActions resultActions = mockMvc.perform(get("/logout"));
+        resultActions.andExpect(status().isFound());
         resultActions.andExpect(redirectedUrl("http://localhost:8080"));
+    }
+
+
+    @Test
+    public void redirectsToReferrerHeaderIfNotLoggedInAndReferrerHeaderExists() throws Exception {
+
+        ResultActions resultActions = mockMvc.perform(get("/logout").header("referer", "https://myApp.coffee"));
+        resultActions.andExpect(status().isFound());
+        resultActions.andExpect(redirectedUrl("https://myApp.coffee"));
     }
 }